+++ zio-http/src/main/scala/zhttp/service/client/experimental/ZClientTest.scala
@@ -1,6 +1,5 @@
 package zhttp.service.client.experimental
 
-//import io.netty.handler.codec.http.{HttpHeaderNames, HttpHeaderValues}
 import zhttp.http.URL.Location
 import zhttp.http._
 import zio.{App, ExitCode, URIO, ZIO}
@@ -13,11 +12,11 @@ object ZClientTest extends App {
   private val PORT = 8081
 
   // pick port automatically.
-  val client       = ZClient.port(PORT) ++ ZClient.threads(2)
+  val client = ZClient.port(PORT) ++ ZClient.threads(2)
 
-//  val keepAliveHeader = Headers(HttpHeaderNames.CONNECTION, HttpHeaderValues.KEEP_ALIVE)
+  //  val keepAliveHeader = Headers(HttpHeaderNames.CONNECTION, HttpHeaderValues.KEEP_ALIVE)
   val emptyHeaders = Headers.empty
-  val req          = ReqParams(
+  val req = ReqParams(
     Method.GET,
     URL(!! / "foo", Location.Absolute(Scheme.HTTP, "localhost", PORT)),
     emptyHeaders,
@@ -37,7 +36,7 @@ object ZClientTest extends App {
   override def run(args: List[String]): URIO[zio.ZEnv, ExitCode] = {
     val resp: ZIO[Any, Throwable, Resp] =
       client
-        .make(req)  // no need to be managed
+        .make(req) // no need to be managed
         .use(_.run)
 
     resp.exitCode
